requirements.txt
Flask==3.1.3
qrcode==8.2
pillow==12.2.0
pdfplumber==0.11.9
PyPDF2==3.0.1
patchright==1.55.2
gunicorn==23.0.0
google-genai>=1.0.0
anthropic>=0.50.0
fpdf2>=2.7.0
lightrag-hku>=1.3.4
